Crate thin_trait_objects

Proof of concept for creating FFI-safe trait objects in Rust.

Structs

FileHandle
A FFI-safe version of the trait object, dyn std::io::Write.

FileHandleBuilder

OwnedFileHandle
An owned wrapper around a *mut FileHandle for use in Rust code.

Functions

file_handle_destroy⚠
Free the FileHandle, calling any destructors and cleaning up any resources being used.

file_handle_flush⚠
Flush this output stream, ensuring that all intermediately buffered contents reach their destination.

file_handle_write⚠
Write some data to the file handle, returning the number of bytes written.

new_file_handle_builder⚠

new_file_handle_from_path⚠
Create a new FileHandle which will write to a file on disk.

new_null_file_handle⚠
Create a new FileHandle which throws away all data written to it.

new_stdout_file_handle⚠
Create a new FileHandle which writes directly to stdout.